+++ tests/test_request.js
@@ -35,6 +35,28 @@ test("request with params", function () {
   same({bar: "baz"}, request.params, "should add any params to the request params object");
 });
 
+test("request with params including new lines", function () {
+  var request = factory('request', {
+    method: 'post',
+    fullPath: '/foo?bar=baz\nzab'
+  })
+
+  equal('/foo', request.path)
+})
+
+test("handling escaped params properly", function () {
+  var escapedPath = '/search?query=foo+bar+%26zoo%3Dgoo&search%5Bbody%5D=herp'
+  var request = factory('request', {
+    method: 'get',
+    fullPath: escapedPath
+  })
+
+  equal('/search', request.path)
+  equal('/search?query=foo bar %26zoo%3Dgoo&search[body]=herp', request.location())
+  equal('foo bar &zoo=goo', request.params.query)
+  equal('herp', request.params.search.body)
+})
+
 test("convert request to readable string", function () {
   var request = factory('request', {
     method: 'post',