+++ CHANGELOG.markdown
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.9.8
+
+* Fix an issue that prevented params with new lines from being handled properly the hash routing plugin, thanks [maxthelion](https://github.com/maxthelion)
+* Fix an issue with the way params are encoded and decoded [issue 70](https://github.com/olivernn/davis.js/issues/70), thanks [Valloric](https://github.com/Valloric)
+
+## 0.9.7
+
+* Pass request options through the location delegate to the location handling object.
+* Pass request options through the request redirect method.
+* Add a simpler hash routing plugin.
+
 ## 0.9.6
 
 * Proper handling of relative paths using $.prop instead of $.attr [issue 58](https://github.com/olivernn/davis.js/issues/58)